FastAPI_Backend/requirements.txt
fastapi== 0.88.0
uvicorn==0.20.0
numpy==1.24.1
scikit-learn==1.1.3
pandas==1.5.1
pydantic==1.10.4
google-generativeai